requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
psycopg2-binary
python-dotenv